Relax errors and improve error messages
In JavaScript, if you try to access a property on an object or an index
of an array that doesn't exist, you get undefined. This commit relaxes
it's validation to better match that behavior. However, it does still
throughs errors if you try to index into a string or other scalar even
though JavaScript allows it. I chose the throw an error here because
it is undefined in JSON Pointer.